% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{create_authorize_url}
\alias{create_authorize_url}
\title{create_authorize_url}
\usage{
create_authorize_url(request_token)
}
\arguments{
\item{request_token}{character string. Pocket request token generated by \code{\link{get_request_token}}.}
}
\value{
character string. The URL for authorizing the Pocket application for which the request token was requested.
}
\description{
creates the URL the user needs to enter into their browser in order to authorize the Pocket application / the request token.
}
\details{
See the \href{https://github.com/CorrelAid/pocketapi}{GitHub README} for details.
}
\seealso{
Other authentication functions: 
\code{\link{get_access_token}()},
\code{\link{get_request_token}()}
}
\concept{authentication functions}
