% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pocket_add.R
\name{extract_action_no_id_results_}
\alias{extract_action_no_id_results_}
\title{extract_action_no_id_results_}
\usage{
extract_action_no_id_results_(res)
}
\arguments{
\item{add_urls}{character vector. URLs that are to be added}

\item{action_name}{character. Name of the action to be used (add)}

\item{...}{additional named arguments to be added to the action list.}
}
\value{
list
}
\description{
generate an action list element for a given action name
}
