% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pocket_modify.R
\name{gen_action}
\alias{gen_action}
\title{gen_action_}
\usage{
gen_action(item_id, action_name)
}
\arguments{
\item{item_id}{character. ID of Pocket item}

\item{action_name}{character. Name of Pocket action as string.}
}
\value{
list
}
\description{
generate an action list element for a given id and action name
}
