% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pocket_modify.R
\name{gen_action_}
\alias{gen_action_}
\title{gen_action_}
\usage{
gen_action_(item_id, action_name, ...)
}
\arguments{
\item{item_id}{character. ID of Pocket item}

\item{action_name}{character. Name of Pocket action as string.}

\item{...}{additional named arguments to be added to the action list.}
}
\value{
list representing a Pocket API action.
}
\description{
generate an Pocket action list element for a given id and action name
}
