% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pocket_add.R
\name{gen_add_action_}
\alias{gen_add_action_}
\title{gen_add_action_}
\usage{
gen_add_action_(add_urls, action_name, ...)
}
\arguments{
\item{add_urls}{Character vector. URLs that are to be added to Pocket.}

\item{action_name}{Character. Name of the action to be used (ADD, in this case).}

\item{...}{Additional named arguments to be added to the action list.}
}
\value{
List of actions and URLs to add to Pocket.
}
\description{
Generate an action list element for adding URLs to Pocket.
}
\keyword{internal}
