% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_access_token}
\alias{get_access_token}
\title{get_access_token}
\usage{
get_access_token(consumer_key, request_token)
}
\arguments{
\item{consumer_key}{character string. Your Pocket consumer key.}

\item{request_token}{character string. Pocket request token generated by \code{\link{get_request_token}}.}
}
\value{
character string. The access token.
}
\description{
creates the URL the user needs to enter into their browser in order to authorize the Pocket application / the request token.
}
\details{
See the \href{https://github.com/CorrelAid/pocketapi}{GitHub README} for details.
}
\seealso{
Other authentication functions: 
\code{\link{create_authorize_url}()},
\code{\link{get_request_token}()}
}
\concept{authentication functions}
