% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_world.R
\name{hello_world}
\alias{hello_world}
\title{hello_world}
\usage{
hello_world(name)
}
\arguments{
\item{name}{character. Name of person / animal / ... to greet.}
}
\value{
character. A nice greeting.
}
\description{
hello_world prints a nice greeting to someone.
}
