% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pocket_get.R
\name{parse_item_}
\alias{parse_item_}
\title{parse_item_}
\usage{
parse_item_(item)
}
\arguments{
\item{item}{Pocket item from the Pocket entry list.}
}
\value{
Tibble
}
\description{
Parse item in the response list into a mini tibble with one row.
}
\keyword{internal}
