% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pocket_add.R
\name{pocket_add}
\alias{pocket_add}
\title{pocket_add}
\usage{
pocket_add(
  add_urls,
  item_ids = "",
  tags = NULL,
  success = TRUE,
  consumer_key = Sys.getenv("POCKET_CONSUMER_KEY"),
  access_token = Sys.getenv("POCKET_ACCESS_TOKEN")
)
}
\arguments{
\item{add_urls}{Character vector. The URL or URLs you want to add to your Pocket list.}

\item{item_ids}{Character vector. (Optional) The item_ids of the items you want to add.}

\item{tags}{Character vector. One or more tags to be applied to all of the newly added URLs.}

\item{success}{Logical. Enables success/failure messages for every URL. Defaults to TRUE. Needs GET access if TRUE.}

\item{consumer_key}{Character string. Your Pocket consumer key. Defaults to \code{Sys.getenv("POCKET_CONSUMER_KEY")}.}

\item{access_token}{Character string. Your Pocket request token. Defaults to \code{Sys.getenv("POCKET_ACCESS_TOKEN")}.}
}
\value{
Invisibly returns the response from the \code{httr} call.
}
\description{
Add one or more items to your Pocket account.
}
