% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pocket_archive.R
\name{pocket_archive}
\alias{pocket_archive}
\title{pocket_archive}
\usage{
pocket_archive(item_ids)
}
\arguments{
\item{item_ids}{character vector. Pocket item ids you want to archive.}

\item{consumer_key}{character. Your Pocket consumer key. Defaults to Sys.getenv("POCKET_CONSUMER_KEY").}

\item{access_token}{character. Your Pocket request token. Defaults to Sys.getenv("POCKET_ACCESS_TOKEN").}
}
\description{
archive items from your Pocket list.
}
