% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pocket_delete.R
\name{pocket_delete}
\alias{pocket_delete}
\title{pocket_delete}
\usage{
pocket_delete(item_ids)
}
\arguments{
\item{item_ids}{character vector. Pocket item ids you want to delete from your list.}

\item{consumer_key}{character. Your Pocket consumer key. Defaults to Sys.getenv("POCKET_CONSUMER_KEY").}

\item{access_token}{character. Your Pocket request token. Defaults to Sys.getenv("POCKET_ACCESS_TOKEN").}
}
\description{
delete items from your Pocket list.
}
