% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pocket_modify.R
\name{pocket_modify_bulk_}
\alias{pocket_modify_bulk_}
\title{pocket_modify_bulk_}
\usage{
pocket_modify_bulk_(item_ids, action_name, consumer_key, access_token, ...)
}
\arguments{
\item{item_ids}{Character vector. Pocket item IDs that should be modified.}

\item{action_name}{Character. The action that should be performed on all specified items.}

\item{consumer_key}{Character string. Your Pocket consumer key. Defaults to \code{Sys.getenv("POCKET_CONSUMER_KEY")}.}

\item{access_token}{Character string. Your Pocket request token. Defaults to \code{Sys.getenv("POCKET_ACCESS_TOKEN")}.}

\item{...}{Additional named arguments to be added to the action list items.}
}
\description{
Bulk modify for a given action, i.e., the action is the same for all item_ids.
}
\keyword{internal}
