% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pocket_tag.R
\name{pocket_tag}
\alias{pocket_tag}
\title{pocket_tag}
\usage{
pocket_tag(item_id = NULL, action_name, tags = NULL, old_new = NULL)
}
\arguments{
\item{item_id}{atomic character vector. Pocket item id you want to modify the tags for.}

\item{action_name}{atomic character vector. The kind of tag action you want to undertake.}

\item{tags}{character vector. The names of the tags to work with the chosen action.}

\item{old_new}{character vector with two elements. Compulsory if action = rename. First element = old tag, second = new tag.}
}
\description{
modify the tags of the items in pocket.
}
