% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_tag_request_}
\alias{process_tag_request_}
\title{process_tag_request_
Validity checks for tag requests}
\usage{
process_tag_request_(item_id, action_name, tags, old_new)
}
\description{
process_tag_request_
Validity checks for tag requests
}
