% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pocket_modify.R
\name{warn_for_failures_}
\alias{warn_for_failures_}
\title{warn_for_failures_}
\usage{
warn_for_failures_(failures)
}
\arguments{
\item{failures}{List of failures.}
}
\description{
Generate warnings for all failures.
}
\keyword{internal}
