% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pocket_add.R
\name{check_for_add_success_}
\alias{check_for_add_success_}
\title{check_for_add_success_}
\usage{
check_for_add_success_(urls)
}
\arguments{
\item{urls}{character vector containing URLs to be checked}

\item{...}{additional named arguments to be added to the action list.}
}
\value{
logical
}
\description{
check whether all urls were successfully added to Pocket
}
