% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{create_authorize_url}
\alias{create_authorize_url}
\title{create_authorize_url}
\usage{
create_authorize_url(request_token)
}
\arguments{
\item{request_token}{Character string. Pocket request token generated by the function \code{\link{get_request_token}}.}
}
\value{
Character string. The URL for authorizing the Pocket application for which the request token was requested. This is what you enter into your browser.
}
\description{
Creates the URL the user needs to enter into her browser in order to authorize the Pocket application/the request token.
}
\details{
See the \href{https://github.com/CorrelAid/pocketapi}{GitHub README} for details on Authentication.
}
\seealso{
Other authentication functions: 
\code{\link{get_access_token}()},
\code{\link{get_request_token}()}
}
\concept{authentication functions}
