% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pocket_modify.R
\name{extract_action_results_}
\alias{extract_action_results_}
\title{extract_action_results_}
\usage{
extract_action_results_(res, item_ids)
}
\arguments{
\item{res}{List. Httr response object.}

\item{item_ids}{Character vector. Pocket item IDs that were modified.}
}
\value{
Named list with the IDs for which the action was successful, the ids for which it failed and the list of failures.
}
\description{
Extract results from list that is returned by the send endpoint of the Pocket API.
}
\keyword{internal}
