% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pocket_modify.R
\name{gen_action_}
\alias{gen_action_}
\title{gen_action_}
\usage{
gen_action_(item_id, action_name, ...)
}
\arguments{
\item{item_id}{Character. ID of Pocket item.}

\item{action_name}{Character. Name of Pocket action as a string.}

\item{...}{Additional, named arguments to be added to the action list.}
}
\value{
List representing a Pocket API action.
}
\description{
Generate a Pocket action list element for a given ID and action name.
}
\keyword{internal}
