% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pocket_add.R
\name{gen_add_action_}
\alias{gen_add_action_}
\title{gen_add_action_}
\usage{
gen_add_action_(add_urls, action_name, ...)
}
\arguments{
\item{add_urls}{character vector. URLs that are to be added}

\item{action_name}{character. Name of the action to be used (add)}

\item{...}{additional named arguments to be added to the action list.}
}
\value{
list
}
\description{
generate an action list element for a given action name
}
