% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pocket_tag.R
\name{gen_tag_action_}
\alias{gen_tag_action_}
\title{gen_tag_action_}
\usage{
gen_tag_action_(action_name, ...)
}
\arguments{
\item{action_name}{Character. Name of Pocket action as a string.}

\item{...}{Additional, named arguments to be added to the action list.}
}
\value{
Action list.
}
\description{
Generate an action list element for a given action name.
}
\keyword{internal}
