% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_access_token}
\alias{get_access_token}
\title{get_access_token}
\usage{
get_access_token(consumer_key, request_token)
}
\arguments{
\item{consumer_key}{Character string. Here goes your Pocket consumer key.}

\item{request_token}{Character string. Pocket request token generated by \code{\link{get_request_token}}.}
}
\value{
Character string. Returns the access token.
}
\description{
Creates the URL the user needs to enter into her browser in order to authorize the Pocket application/the request token.
}
\details{
See the \href{https://github.com/CorrelAid/pocketapi}{GitHub README} for details on Authentication.
}
\seealso{
Other authentication functions: 
\code{\link{create_authorize_url}()},
\code{\link{get_request_token}()}
}
\concept{authentication functions}
