% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pocket_add.R
\name{pocket_add}
\alias{pocket_add}
\title{pocket_add}
\usage{
pocket_add(
  add_urls,
  item_ids = "",
  tags = NULL,
  success = TRUE,
  consumer_key = Sys.getenv("POCKET_CONSUMER_KEY"),
  access_token = Sys.getenv("POCKET_ACCESS_TOKEN")
)
}
\arguments{
\item{add_urls}{character vector. The URL or URLs you want to add to your Pocket list.}

\item{item_ids}{character vector. (Optional) The item_ids of the items you want to add.}

\item{tags}{character vector. One or more tags to be applied to any of the newly added URLs.}

\item{success}{logical. Enables success/failure messages for each URL. Defaults to TRUE.}

\item{consumer_key}{character string. Your Pocket consumer key. Defaults to Sys.getenv("POCKET_CONSUMER_KEY").}

\item{access_token}{character string. Your Pocket request token. Defaults to Sys.getenv("POCKET_ACCESS_TOKEN").}
}
\value{
the response from the httr call, invisibly
}
\description{
add one or more items to Pocket
}
