% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pocket_add.R
\name{pocket_add}
\alias{pocket_add}
\title{pocket_add}
\usage{
pocket_add(consumer_key, access_token, add_url)
}
\arguments{
\item{consumer_key}{character string. Your Pocket consumer key. See https://getpocket.com/developer/docs/authentication.}

\item{access_token}{character string. Your Pocket request token. See https://getpocket.com/developer/docs/authentication.}

\item{add_url}{character string. The URL of the item you want to add to your Pocket list.}
}
\description{
add entry to pocket
}
