% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pocket_get.R
\name{pocket_get}
\alias{pocket_get}
\title{pocket_get}
\usage{
pocket_get(
  favorite = NULL,
  item_type = NULL,
  tag = NULL,
  state = "all",
  consumer_key = Sys.getenv("POCKET_CONSUMER_KEY"),
  access_token = Sys.getenv("POCKET_ACCESS_TOKEN")
)
}
\arguments{
\item{favorite}{Boolean. Defaults to \code{NULL}. Allows to filter for favorited items. If TRUE, only favorited items will be returned. If FALSE, only un-favorited items will be returned.}

\item{item_type}{Character. Defaults to \code{NULL}. Allows to filter for content type of items. Valid values are: "image", "article", "video". Please note that there might be Pocket items that do not belong to any of those types. The Pocket API documentation only mentions those three.}

\item{tag}{Character. Defaults to \Code{NULL}. Only one tag can be filtered for at a time. Set to "_untagged_" if you only want to get untagged items.}

\item{state}{Character. Defaults to "all". Allows to filter on unread/archived items or return all. Valid values are "unread", "archive", "all".}

\item{consumer_key}{Character string. Your Pocket consumer key. Defaults to \code{Sys.getenv("POCKET_CONSUMER_KEY")}.}

\item{access_token}{Character string. Your Pocket request token. Defaults to \code{Sys.getenv("POCKET_ACCESS_TOKEN")}.}
}
\value{
Tibble with one row for each Pocket item.
}
\description{
Get a data frame with your pocket data, including item IDs.
}
\details{
See https://getpocket.com/developer/docs/v3/retrieve for the meaning of the variable values in the resulting tibble.
}
