% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pocket_get.R
\name{pocket_get}
\alias{pocket_get}
\title{pocket_get2}
\usage{
pocket_get(
  consumer_key = Sys.getenv("POCKET_CONSUMER_KEY"),
  access_token = Sys.getenv("POCKET_ACCESS_TOKEN")
)
}
\arguments{
\item{consumer_key}{character. Your Pocket consumer key. Defaults to Sys.getenv("POCKET_CONSUMER_KEY").}

\item{access_token}{character. Your Pocket request token. Defaults to Sys.getenv("POCKET_ACCESS_TOKEN").}
}
\value{
tibble. Tibble with one row for each Pocket item.
}
\description{
Get a data frame with your pocket data.
}
\details{
See https://getpocket.com/developer/docs/v3/retrieve for the meaning of certain variable values.
}
