% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pocket_modify.R
\name{pocket_modify}
\alias{pocket_modify}
\title{pocket_modify}
\usage{
pocket_modify(actions)
}
\arguments{
\item{actions}{list. list of lists where each element is an action object. See https://getpocket.com/developer/docs/v3/modify.}
}
\description{
internal function that sends a request with actions to the modify Pocket API endpoint.
}
\details{
see https://getpocket.com/developer/docs/v3/modify.
}
