% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pocket_unfavorite.R
\name{pocket_unfavorite}
\alias{pocket_unfavorite}
\title{pocket_unfavorite}
\usage{
pocket_unfavorite(
  item_ids,
  consumer_key = Sys.getenv("POCKET_CONSUMER_KEY"),
  access_token = Sys.getenv("POCKET_ACCESS_TOKEN")
)
}
\arguments{
\item{item_ids}{character vector. Pocket item ids you want to unfavorite}

\item{consumer_key}{character. Your Pocket consumer key. Defaults to Sys.getenv("POCKET_CONSUMER_KEY").}

\item{access_token}{character. Your Pocket request token. Defaults to Sys.getenv("POCKET_ACCESS_TOKEN").}
}
\description{
unfavorite items from your Pocket list.
}
